# docs/swagger.yaml
basePath: /
definitions:
  controller.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dtos.BookCreate:
    properties:
      author:
        type: string
      description:
        type: string
      pages:
        type: integer
      title:
        type: string
    type: object
  dtos.BookPatch:
    properties:
      author:
        type: string
      description:
        type: string
      pages:
        type: integer
      title:
        type: string
    type: object
  dtos.BookUpdate:
    properties:
      author:
        type: string
      description:
        type: string
      pages:
        type: integer
      title:
        type: string
    type: object
  dtos.BookWithAuthor:
    properties:
      author:
        $ref: '#/definitions/dtos.ResponseAuthor'
      book:
        $ref: '#/definitions/entity.Book'
    type: object
...
  dtos.ResponseAuthor:
    properties:
      birthday:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  entity.Book:
    properties:
      author:
        type: string
      description:
        type: string
      id:
        type: string
      pages:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: gui.rialli@gmail.com
    name: Guilherme Rialli
    url: www.github.com/guirialli/
  description: rater limit example
  title: Rater Limit
  version: "1.0"
paths:
...
  /books:
    get:
      description: Retrieve a list of all books
      produces:
      - application/json
      responses:
        "200":
          description: List of books
          schema:
            items:
              $ref: '#/definitions/entity.Book'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Add a new book to the database
      parameters:
      - description: Book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dtos.BookCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created book
          schema:
            $ref: '#/definitions/entity.Book'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      description: Remove a book from the database
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a book
      tags:
      - books
    get:
      description: Retrieve a specific book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book details
          schema:
            $ref: '#/definitions/entity.Book'
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get book by ID
      tags:
      - books
    patch:
      consumes:
      - application/json
      description: Partially update book details in the database
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Patch data for the book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dtos.BookPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Patched book
          schema:
            $ref: '#/definitions/entity.Book'
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Patch an existing book
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update book details in the database
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dtos.BookUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Updated book
          schema:
            $ref: '#/definitions/entity.Book'
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing book
      tags:
      - books
  /books/{id}/author:
    get:
      description: Retrieve a specific book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book details
          schema:
            $ref: '#/definitions/dtos.BookWithAuthor'
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get book by ID with Author
      tags:
      - books
  /books/author:
    get:
      description: Retrieve a list of all books along with their authors
      produces:
      - application/json
      responses:
        "200":
          description: List of books with authors
          schema:
            items:
              $ref: '#/definitions/dtos.BookWithAuthor'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all books with authors
      tags:
      - books
...
  /public/books:
    get:
      description: Retrieve a list of all books
      produces:
      - application/json
      responses:
        "200":
          description: List of books
          schema:
            items:
              $ref: '#/definitions/entity.Book'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all books
      tags:
      - books
  /public/books/{id}:
    get:
      description: Retrieve a specific book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book details
          schema:
            $ref: '#/definitions/dtos.BookWithAuthor'
        "400":
          description: ID is required
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get book by ID with Author
      tags:
      - books
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"